ref-filter: add multiple-option parsing functions
The functions

	match_placeholder_arg_value()
	match_placeholder_bool_arg()

were added in pretty (4f732e0fd7 (pretty: allow %(trailers) options
with explicit value, 2019-01-29)) to parse multiple options in an
argument to --pretty. For example,

	git log --pretty="%(trailers:key=Signed-Off-By,separator=%x2C )"

will output all the trailers matching the key and seperates them by
commas per commit.

Add similar functions,

	match_atom_arg_value()
	match_atom_bool_arg()

in ref-filter. A particular use of this can be seen in the subsequent
commit where we parse the options given to a new atom "describe".

Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Hariom Verma <hariom18599@gmail.com>
Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>