ref-filter: add mailmap support
Add mailmap support to ref-filter to formats which are similar in
pretty. This support is such that the following pretty placeholders are
equivalent to the new ref-filter atoms

	%aN = authorname:mailmap
	%cN = committername:mailmap

	%aE = authoremail:mailmap
	%aL = authoremail:mailmap,localpart    (this order matters)
	%cE = committeremail:mailmap
	%cL = committeremail:mailmap,localpart (this order matters)

Additionally, mailmap can also be used with ":trim" option for email by
doing something like "authoremail:mailmap,trim" (this order matters).

The above also applies for the "tagger" atom, that is,
"taggername:mailmap", "taggeremail:mailmap", "taggeremail:mailmap,trim"
and "taggername:mailmap,localpart" (the order of options matters here
again).

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>