find_wholine: return the entire buffer
Make the function find_wholine(), in ref-filter, return the entire
buffer instead of stripping off the "who" part.

This is useful when we are calling functions on this buffer which would
like to know the "who" part of the buffer so that they can do something
with it.

This is especially useful in the following commit.

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>