pretty: use ref-filter logic wherever possible
Whenever we have a placeholder that translates to a ref format, use
ref-filter logic to format the output instead of the usual pretty logic.

When such a translation is not found, fall back to pretty logic.

This takes us one step closer to unifying both the ways of doing things.

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>