ref-filter: introduce get_mailmap_buffer()
If the ":mailmap" option is used, while the grabbing the values for the
corresponding atom, we change the idents in the buffer to match with the
idents in the .mailmap file (or according to the file set in the
"mailmap.file" or "mailmap.blob" config variables).

Currently, this transforming of buf is done directly in grab_person().
Instead, do it in a new function "get_mailmap_buffer()" so that this
functionality can be reused elsewhere in ref-filter (where ever we need
to show the mailmap values).

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>