ref-filter: add "reflog" atom
Add new "reflog" atom to ref-filter to support viewing of reflog info of
refs.

This also adds to the effort of duplicating the formats supported by
pretty in ref-filter.  The reflog formats in pretty correspond to the
atoms in ref-filter in the following manner:

	%gd = %(reflog)
	%gD = %(reflog:short)
	%gn = %(reflog:name)
	%gN = %(reflog:name,mailmap)
	%ge = %(reflog:email)
	%gE = %(reflog:email,mailmap)
	%gs = %(reflog:subject)

It is also possible to add a couple more options in the context of the
email options - "trim" and "localpart" - and any kind of a combination of
them with or without "mailmap" (like "trim,localpart,mailmap"), as these
are the options that are supported by "{author,committer,tagger}email"
atoms, but the current implementation doesn't support such features.

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>