ref-filter: add "reflog" atom
Add new "reflog" atom to ref-filter to support viewing of reflog info of
refs.

This also adds to the effort of duplicating the formats supported by
pretty in ref-filter.  The reflog formats in pretty correspond to the
atoms in ref-filter in the following manner:

	%gd = %(reflog)
	%gD = %(reflog:short)
	%gn = %(reflog:name)
	%ge = %(reflog:email)
	%gs = %(reflog:subject)

It is also possible to add a few more options in the context of the
email options - "trim", "localpart" and "mailmap"- and any kind of a
combination of them (like "trim,localpart,mailmap"), as these are the
options that are supported by "{author,committer,tagger}email" atoms,
but the current implementation doesn't support such features.

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>