ident: introduce apply_mailmap_to_ident_line()
Add a new function apply_mailmap_to_ident_line() to change lines of the
form

			name <email>

to their mailmap equivalents.

This function is a convenience wrapper around the private function
"rewrite_ident_line()".

It is intended to be used in cases where our ident is of the form as
mentioned above and an example usage of this can be found in the
subsequent commit.

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>