pretty: refactor match_placeholder_*() functions
f46094a (ref-filter: add multiple-option parsing functions, 2023-07-24)
introduced the match_atom_*() functions in ref-filter, which mimicked
what match_placeholder_*() functions did, but for ref-filter atoms.  This
was done because of the nature of input given to these functions.

  match_placeholder*() expect an input string which terminates with ')'[1]
  match_atom*() expect an input string which terminates with '\0' ie NUL

Therefore, refactor match_placeholder_*() so that they can handle being
given an input string which terminates at NUL or at ')', so that they
can also be used in ref-filter code.

[1]
NOTE:
One should not be lead to beleive that the current match_placeholder_*()
functions are not NUL-terminated strings.  They are.  Although the
functions see the ')' character and understand that they must return.

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>