pretty: refactor match_placeholder_*() functions
f46094a (ref-filter: add multiple-option parsing functions, 2023-07-24)
introduced the match_atom_*() functions in ref-filter, which mimicked
what match_placeholder_*() functions did, but for ref-filter atoms.  This
was done because of the nature of input given to these functions -

  match_placeholder*() expect an input string which terminates with ')'
	so something like "abbrev=10,tags)"

  match_atom*() expect an input string which terminates with NUL
	so something like "abbrev=10,tags"

  when the given format to both pretty and ref-filter (ie say through
  log and for-each-ref respectively) is "%(describe:abbrev=10,tags)".

Therefore, refactor match_placeholder_*() so that they can handle being
given an input string which terminates at ')' or at NUL - whichever
occurs first - so that they can also be used in ref-filter code.

In addition to this refactoring,

- drop the NULL check for "valuestart" within the implementation of
  "match_placheholder_arg_value()" since it is useless, as callers
  already pass a non NULL "valuestart" and requiring them to pass a non
  NULL "valuestart" is a simple enough requirement.

- don't allow for parsing when there are '('s in the input.  ref-filter
  doesn't work well with parentheses in the formats passed to it.  We
  can't really handle ')'s as they get stripped off much higher in the
  call chain and a fix there would be more appropriate than here.

  This in turn means that we should also remove this check for '(' when
  ref-filter learns to work well with parentheses in general.

Sidenote:  It might not be obvious but the existing match_placeholder_*()
functions are always passed NUL-terminated strings.  Although the
functions see the ')' character and understand that they must return.
This holds true for the current refactoring too.  That is, if ')'
character appears first, then the function will return immediately and if
not then will look for NUL [1].

Finally, again, all of this is because the functions in pretty which are
responsible for parsing the format expect their argument to end with
')', while those in ref-filter expect their argument to end with NUL.
Hence the refactoring focuses on only these two characters and the
functions now expect the input string to be in the respective manner.

[1]:  This statement is only partially true because it is possible to
have a situation where there are ')' s embedded in the "candidate" that
we have at the beginning of "to_parse", such as

	to_parse="ke)y=value"
	candidate="ke)y"

in which case we just skip over it and the function returns successfully.

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>