ref-filter: introduce match_format_*() functions
The match_placeholder_*() functions were refactored in the previous
commit so that they can now be used elsewhere outside of pretty.  To
allow for doing so, rename these functions to a more general name

	match_format_arg_value()
	match_format_bool_arg()

While at it, move them to ref-filter and make them visible globally.

The move to ref-filter is because, at some point in the future we intend
to have a single interface for formats and this is also the reason why
there has been effort in the past to unify ref-filter formats and the
pretty formats, so that ref-filter can support all the formats that
pretty supports.  Since we intend for everthing to be present in
ref-filter in the end, it makes sense that the match_format_*()
functions are present in ref-filter.

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>