pretty: introduce match_format_arg_value() and match_format_bool_arg()
f46094a (ref-filter: add multiple-option parsing functions, 2023-07-24)
introduced the match_atom_*() functions in ref-filter, which mimicked
what match_placeholder_*() functions did, but for ref-filter atoms.  This
was done because of the nature of input given to these functions -

  match_placeholder*() expect an input string which terminates with ')'
        so something like "abbrev=10,tags)"

  match_atom*() expect an input string which terminates with NUL
        so something like "abbrev=10,tags"

  when the given format to both pretty and ref-filter (ie say through
  log and for-each-ref respectively) is "%(describe:abbrev=10,tags)".

Therefore, refactor match_placeholder_arg_value() so that it can handle
being given any terminating character in its "to_parse".

In addition to this refactoring, drop the NULL check for "valuestart"
within the implementation of "match_placheholder_arg_value()" since it
is useless, as callers already pass a non NULL "valuestart" and requiring
them to pass a non NULL "valuestart" is a simple enough requirement.

If the caller is sure that they don't have a value then they are
better off using skip_prefix() instead of this function.

While at it, rename it to "match_format_arg_value()" addressing its
generality and rename "match_placeholder_bool_arg()" to
"match_format_bool_arg()" respectively.

The following commit will move this function to ref-filter and make it
globally visible.

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>