ref-filter: make match_format_arg_value() global
Move the "match_format_arg_value()" function to ref-filter and make it
global.  Hence also wrap the "match_atom_*()" functions around this
function.

The move to ref-filter is because, at some point in the future we intend
to have a single interface for formats and this is also the reason why
there has been effort in the past to unify ref-filter formats and the
pretty formats, so that ref-filter can support all the formats that
pretty supports.  Since we intend for everything to be present in
ref-filter in the end, it makes sense that the match_format_arg_value()
is present in ref-filter.

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>