t6300: values containing ')' are broken in ref formats
Document that values containing ')' in formats are not parsed correctly
in ref-filter.

The problem here is that ref-filter, while parsing the format string,
looks for the first occurence of ')' and marks it to be the end of _that_
particular atom - which is obviously wrong in cases where the format is
of type

	atom:key=value

where "value" has ')' somewhere in it.

For example, in %(if:equals=somere)f), the string that ref-filter should
compare against is "somere)f", although since the parsing behavior in
these cases is broken, we instead compare against "somere".

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>