ref-filter: make match_format_*() functions global
Move "match_format_arg_value()" and "match_format_bool_arg()" to
ref-filter and make them global, hence also removing match_atom_*()
functions and replacing their calls with the newly introduced functions.

The move to ref-filter is because, at some point in the future we intend
to have a single interface for formats and this is also the reason why
there has been effort in the past to unify ref-filter formats and the
pretty formats, so that ref-filter can support all the formats that
pretty supports.  Since we intend for everything to be present in
ref-filter in the end, it makes sense that "match_format_arg_value()"
and "match_format_bool_arg()" are present in ref-filter too.

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>